Update: Major update across memory, knowledge, multimodal, core, concurrency, GPU, training, frontend, and utils
- Memory: memory (h/c), memory_manager (h/c), semantic, short_term, working_memory
- Knowledge: knowledge.h, graph_query, graph_reasoning, graph_storage, knowledge_graph, knowledge_inference, knowledge_integration, knowledge_self_check, semantic_network, skill_library, uncertainty_reasoning
- Multimodal: audio_semantic, data_collection_pipeline, motor, multimodal_integration, multimodal_unified_input (h/c), proprioception, radar, sensor, slam (h/c internal), slam_enhance, speech_recognition, text, thermal
- Core: lnn.h, unified_lnn_state (h/c), cfc_cell, cfc_network, decision_engine, extended_training, laplace_unified, lnn, quaternion_conv, selflnn, state, tensor, CMakeLists; Deleted safe_memory.h
- AGI: agi, capability_switch, CMakeLists
- Concurrency: lock_free, rcu, rw_lock_map
- GPU: gpu_ascend, gpu_cambricon, gpu_intel, gpu_metal, gpu_rocm, gpu_tpu, npu, npu_common
- Backend: backend, websocket_push
- Training: data_loaders (h/c), training, training_data_pipeline, training_dataset, training_monitor (h/c), mixed_precision, regularization (h/c), training_pipeline; Deleted dataset_api.h
- Evolution: evolution_engine, neural_architecture_search
- Reasoning: cfc_uncertainty_reasoning, reasoning_internal
- Learning: imitation_learning, meta_learning, online_learning, reinforcement_learning
- Cognition: deep_thought_chain, metacognition
- Robot: dynamics, kinematics, ros2_node
- Utils: json_parser, memory_utils, secure_random; Deleted deep_copy_utils.h
- Safety: (implicit)
- Frontend: style.css, index.html, agi-controller, data-engine, device-manager, dialogue-enhanced, knowledge-graph, main, product-design, selflnn-chart, simulation-control, training-center, training-push, visualization, voice-command
- selflnn.h, CMakeLists.txt, README.md, main.c

+++ src/utils/secure_random.c
@@ -53,7 +53,8 @@ float secure_random_float(void)
 {
     uint32_t r = 0;
     secure_random_bytes((uint8_t*)&r, sizeof(r));
-    return (float)(r / 4294967296.0f);
+    /* ZSFX-DEEP-R12-002: 使用double中间精度防止r=2^32-1时float上溢到1.0 */
+    return (float)((double)r / 4294967296.0);
 }
 
 uint32_t secure_random_int(uint32_t max)
